+++ mod-source-record-manager-server/src/main/java/org/folio/verticle/consumers/StoredRecordChunksKafkaHandler.java
@@ -1,5 +1,6 @@
 package org.folio.verticle.consumers;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.Json;
@@ -14,12 +15,15 @@ import org.folio.kafka.AsyncRecordHandler;
 import org.folio.kafka.KafkaHeaderUtils;
 import org.folio.rest.jaxrs.model.DataImportEventTypes;
 import org.folio.rest.jaxrs.model.Event;
+import org.folio.rest.jaxrs.model.JobExecution;
 import org.folio.rest.jaxrs.model.JournalRecord;
 import org.folio.rest.jaxrs.model.JournalRecord.EntityType;
+import org.folio.rest.jaxrs.model.ProfileSnapshotWrapper;
 import org.folio.rest.jaxrs.model.Record;
 import org.folio.rest.jaxrs.model.Record.RecordType;
 import org.folio.rest.jaxrs.model.RecordsBatchResponse;
 import org.folio.services.EventProcessedService;
+import org.folio.services.JobExecutionService;
 import org.folio.services.MappingRuleCache;
 import org.folio.services.RecordsPublishingService;
 import org.folio.services.entity.MappingRuleCacheKey;
@@ -29,14 +33,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
+import javax.ws.rs.NotFoundException;
 import java.util.Date;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static java.lang.String.format;
 import static org.apache.commons.lang3.ObjectUtils.allNotNull;
 import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_EDIFACT_RECORD_CREATED;
+import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_MARC_BIB_FOR_ORDER_CREATED;
 import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_SRS_MARC_AUTHORITY_RECORD_CREATED;
 import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_SRS_MARC_BIB_RECORD_CREATED;
 import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_SRS_MARC_HOLDING_RECORD_CREATED;
@@ -53,6 +61,10 @@ public class StoredRecordChunksKafkaHandler implements AsyncRecordHandler<String
   private static final Logger LOGGER = LogManager.getLogger();
   private static final String INSTANCE_TITLE_FIELD_PATH = "title";
   public static final String STORED_RECORD_CHUNKS_KAFKA_HANDLER_UUID = "4d39ced7-9b67-4bdc-b232-343dbb5b8cef";
+  public static final String ORDER_TYPE = "ORDER";
+  public static final String FOLIO_RECORD = "folioRecord";
+  public static final String ACTION_FIELD = "action";
+  public static final String CREATE_ACTION = "CREATE";
 
   private static final Map<RecordType, DataImportEventTypes> RECORD_TYPE_TO_EVENT_TYPE = Map.of(
     MARC_BIB, DI_SRS_MARC_BIB_RECORD_CREATED,
@@ -65,17 +77,20 @@ public class StoredRecordChunksKafkaHandler implements AsyncRecordHandler<String
   private EventProcessedService eventProcessedService;
   private JournalService journalService;
   private MappingRuleCache mappingRuleCache;
+  private JobExecutionService jobExecutionService;
   private Vertx vertx;
 
   public StoredRecordChunksKafkaHandler(@Autowired @Qualifier("recordsPublishingService") RecordsPublishingService recordsPublishingService,
                                         @Autowired @Qualifier("journalServiceProxy") JournalService journalService,
                                         @Autowired @Qualifier("eventProcessedService") EventProcessedService eventProcessedService,
+                                        @Autowired JobExecutionService jobExecutionService,
                                         @Autowired MappingRuleCache mappingRuleCache,
                                         @Autowired Vertx vertx) {
     this.recordsPublishingService = recordsPublishingService;
     this.eventProcessedService = eventProcessedService;
     this.journalService = journalService;
     this.mappingRuleCache = mappingRuleCache;
+    this.jobExecutionService = jobExecutionService;
     this.vertx = vertx;
   }
 
@@ -102,8 +117,18 @@ public class StoredRecordChunksKafkaHandler implements AsyncRecordHandler<String
 
           LOGGER.debug("handle:: RecordsBatchResponse has been received, starting processing chunkId: {} chunkNumber: {} jobExecutionId: {}", chunkId, chunkNumber, jobExecutionId);
           saveCreatedRecordsInfoToDataImportLog(storedRecords, okapiConnectionParams.getTenantId());
-          return recordsPublishingService.sendEventsWithRecords(storedRecords, jobExecutionId,
-              okapiConnectionParams, eventType.value())
+          return jobExecutionService.getJobExecutionById(jobExecutionId, okapiConnectionParams.getTenantId())
+            .compose(jobExecutionOptional -> {
+              if (jobExecutionOptional.isPresent()) {
+                LOGGER.debug("handle:: JobExecution found by id {}: chunkId:{} chunkNumber: {} ", jobExecutionId, chunkId, chunkNumber);
+                return setOrderEventTypeIfNeeded(jobExecutionOptional.get(), eventType);
+              } else {
+                LOGGER.warn("handle:: Couldn't find JobExecution by id {}: chunkId:{} chunkNumber: {} ", jobExecutionId, chunkId, chunkNumber);
+                return Future.failedFuture(new NotFoundException(format("Couldn't find JobExecution with id %s chunkId:%s chunkNumber: %s", jobExecutionId, chunkId, chunkNumber)));
+              }
+            })
+            .compose(eventTypes -> recordsPublishingService.sendEventsWithRecords(storedRecords, jobExecutionId,
+              okapiConnectionParams, eventTypes.value()))
             .compose(b -> {
               LOGGER.debug("handle:: RecordsBatchResponse processing has been completed chunkId: {} chunkNumber: {} jobExecutionId: {}", chunkId, chunkNumber, jobExecutionId);
               return Future.succeededFuture(chunkId);
@@ -116,6 +141,33 @@ public class StoredRecordChunksKafkaHandler implements AsyncRecordHandler<String
       LOGGER.warn("handle:: Can't process kafka record: ", e);
       return Future.failedFuture(e);
     }
+  }
+
+  private Future<DataImportEventTypes> setOrderEventTypeIfNeeded(JobExecution jobExecution, DataImportEventTypes dataImportEventTypes) {
+    if (jobExecution.getJobProfileSnapshotWrapper() != null) {
+      ProfileSnapshotWrapper profileSnapshotWrapper = new ObjectMapper().convertValue(jobExecution.getJobProfileSnapshotWrapper(), ProfileSnapshotWrapper.class);
+      List<ProfileSnapshotWrapper> actionProfiles = profileSnapshotWrapper
+        .getChildSnapshotWrappers()
+        .stream()
+        .filter(e -> e.getContentType() == ProfileSnapshotWrapper.ContentType.ACTION_PROFILE)
+        .collect(Collectors.toList());
+
+      if (!actionProfiles.isEmpty() && checkIfOrderCreateActionProfileExists(actionProfiles)) {
+        dataImportEventTypes = DI_MARC_BIB_FOR_ORDER_CREATED;
+        LOGGER.debug("setOrderEventTypeIfNeeded:: Event type for Order's logic set by jobExecutionId {} ", jobExecution.getId());
+      }
+    }
+    return Future.succeededFuture(dataImportEventTypes);
+  }
+
+  private static boolean checkIfOrderCreateActionProfileExists(List<ProfileSnapshotWrapper> actionProfiles) {
+    for (ProfileSnapshotWrapper actionProfile : actionProfiles) {
+      LinkedHashMap<String, String> content = new ObjectMapper().convertValue(actionProfile.getContent(), LinkedHashMap.class);
+      if (content.get(FOLIO_RECORD).equals(ORDER_TYPE) && content.get(ACTION_FIELD).equals(CREATE_ACTION)) {
+        return true;
+      }
+    }
+    return false;
   }
 
   private void saveCreatedRecordsInfoToDataImportLog(List<Record> storedRecords, String tenantId) {